# backend/src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Medilink Healthcare API",
    "description": "Comprehensive Interactive API Documentation for Medilink Healthcare Platform",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local Development Server"
    },
    {
      "url": "/",
      "description": "Relative server route"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT authentication token."
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/payment/verify-qr": {
      "post": {
        "summary": "Verify QR Payment",
        "tags": ["Payment"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["qrId"],
                "properties": {
                  "qrId": { "type": "string", "example": "qr_Hsp19283" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
    "/api/payment/mark-cash": {
      "post": {
        "summary": "Record Cash Payment",
        "tags": ["Payment"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["referenceId"],
                "properties": {
                  "referenceId": { "type": "string", "format": "uuid", "example": "cc0c4e12-3b2d-4f1a-8c5e-9a1b2c3d4e5f" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
...
  }
}